+++ docs/components/DataList/Web.stories.tsx
@@ -18,8 +18,7 @@ import { Chip } from "@jobber/components/Chip";
 import { Icon } from "@jobber/components/Icon";
 import { Combobox, type ComboboxOption } from "@jobber/components/Combobox";
 import { Flex } from "@jobber/components/Flex";
-// eslint-disable-next-line import/no-internal-modules
-import { useDebounce } from "@jobber/components/utils/useDebounce";
+import { useDebounce } from "@jobber/hooks/useDebounce";
 import { isNormalClick } from "@jobber/components";
 
 const meta: Meta = {