+++ packages/components-native/src/Typography/index.ts
@@ -17,6 +17,7 @@ export type {
   TextVariation,
   TypographyProps,
   TruncateLength,
+  OnTextLayoutEvent,
 } from "./Typography";
 export { getTypographyStyles, useTypographyStyles } from "./Typography.style";
 export { typographyStyles } from "./Typography.style";