+++ packages/components/src/Autocomplete/Autocomplete.tsx
@@ -7,6 +7,7 @@ import React, {
   useRef,
   useState,
 } from "react";
+import { useDebounce } from "@jobber/hooks/useDebounce";
 import styles from "./Autocomplete.module.css";
 import { Menu } from "./Menu/Menu";
 import {
@@ -16,7 +17,6 @@ import {
 } from "./Autocomplete.types";
 import { isOptionGroup } from "./Autocomplete.utils";
 import { InputText, InputTextRef } from "../InputText";
-import { useDebounce } from "../utils/useDebounce";
 import { mergeRefs } from "../utils/mergeRefs";
 
 // Max statements increased to make room for the debounce functions