+++ packages/components/src/Banner/Banner.types.tsx
@@ -1,1 +1,0 @@
-export type BannerType = "notice" | "success" | "warning" | "error";