+++ packages/components/src/Chips/InternalChipDismissible/hooks/useInternalChipDismissible.ts
@@ -1,4 +1,4 @@
-import { KeyboardEvent, MouseEvent, useRef } from "react";
+import { KeyboardEvent, MouseEvent, useState } from "react";
 import sortBy from "lodash/sortBy";
 import { useLiveAnnounce } from "@jobber/hooks/useLiveAnnounce";
 import { InternalChipDismissibleProps } from "../InternalChipDismissibleTypes";
@@ -38,7 +38,9 @@ export function useInternalChipDismissible({
   onClick,
   onCustomAdd,
 }: InternalChipDismissibleProps) {
-  const ref = useRef<HTMLDivElement>(null);
+  const [wrapperElement, setWrapperElement] = useState<HTMLDivElement | null>(
+    null,
+  );
   const chipOptions = children.map(chip => chip.props);
   const visibleChipOptions = chipOptions.filter(chip =>
     selected.includes(chip.value),
@@ -120,7 +122,8 @@ export function useInternalChipDismissible({
 
   return {
     ...actions,
-    ref,
+    ref: setWrapperElement,
+    wrapperElement,
     sortedVisibleChipOptions,
     availableChipOptions,
   };